Refactor page content into content modules and section components
