+++ client/src/pages/Verstehen.tsx
@@ -1,86 +1,19 @@
-import { useState, useRef } from "react";
+import { useState } from "react";
 import SEO from "@/components/SEO";
 import Layout from "@/components/Layout";
 import { Card, CardContent } from "@/components/ui/card";
 import { motion } from "framer-motion";
-import { BookOpen, Brain, Heart, AlertCircle, Lightbulb, ArrowRight, ExternalLink, FileText, Clock, RefreshCw, Layers, Users, Activity, XCircle, Download, Filter } from "lucide-react";
+import { BookOpen, Brain, Heart, AlertCircle, Lightbulb, ArrowRight, ExternalLink, FileText, Clock, RefreshCw, Layers, Users, Activity, XCircle } from "lucide-react";
 import { Link } from "wouter";
 import { Button } from "@/components/ui/button";
 import { TableOfContents } from "@/components/UXEnhancements";
 import ContentSection from "@/components/ContentSection";
 import MythosFlipCards from "@/components/interactive/MythosFlipCards";
 import EvidenceNote from "@/components/EvidenceNote";
+import VerstehenInfografikenSection from "@/sections/VerstehenInfografikenSection";
 
-
-const verstehenInfografiken = [
-  {
-    id: "leuchtturm",
-    title: "Der Leuchtturm",
-    description: "Ihre Rolle als Angehörige/r: Stabil bleiben trotz Sturm.",
-    category: "grundlagen",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/GbFCyQhEWIKomzXw.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/DNGijMOYFghXAsLm.pdf",
-    alt: "Der Leuchtturm – Ihre Rolle als Angehörige/r",
-    featured: true,
-  },
-  {
-    id: "eisberg",
-    title: "Der Eisberg",
-    description: "Wut ist oft nur die Spitze – darunter liegen Schmerz und Angst.",
-    category: "grundlagen",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/MLLwefeyaKvtThbK.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/RNKtfQQMvhlSyiIp.pdf",
-    alt: "Der Eisberg – Wut ist oft die Spitze",
-  },
-  {
-    id: "spaltung",
-    title: "Spaltung",
-    description: "Das Pendel zwischen Extremen – die Grauzone stärken.",
-    category: "grundlagen",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/WRORriPmZftmvKTL.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/RtdVgflJuCNAhEKk.pdf",
-    alt: "Spaltung – das Pendel zwischen Extremen",
-  },
-  {
-    id: "alarm-modus",
-    title: "Alarm-Modus vs. Denk-Modus",
-    description: "Erst beruhigen, dann klären – warum Logik manchmal nicht ankommt.",
-    category: "neurobiologie",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/sSUJoOUTiuWgrkiZ.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/tlKAOpYHdCNAtovE.pdf",
-    alt: "Alarm-Modus vs. Denk-Modus",
-  },
-  {
-    id: "4-phasen",
-    title: "Der 4-Phasen-Zyklus",
-    description: "Das vorhersehbare Muster – Krisen folgen oft einem Ablauf.",
-    category: "neurobiologie",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/BYDbBJaIhetrjHRq.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/PBYpNxZamAxjOHYd.pdf",
-    alt: "Der 4-Phasen-Zyklus",
-  },
-  {
-    id: "gehirn",
-    title: "Das Gehirn verstehen",
-    description: "Neurobiologie einfach erklärt – warum Stress Denken blockiert.",
-    category: "neurobiologie",
-    webpUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/ImASzOTHYdFpxOUI.webp",
-    pdfUrl: "https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/NViSBQtRBvGWOHPE.pdf",
-    alt: "Das Gehirn verstehen",
-  },
-];
-
-const verstehenCategories = [
-  { id: "alle", label: "Alle", icon: Filter, count: verstehenInfografiken.length },
-  { id: "grundlagen", label: "Grundlagen", icon: BookOpen, count: verstehenInfografiken.filter(i => i.category === "grundlagen").length },
-  { id: "neurobiologie", label: "Neurobiologie", icon: Brain, count: verstehenInfografiken.filter(i => i.category === "neurobiologie").length },
-];
 
 export default function Verstehen() {
-  const [activeFilter, setActiveFilter] = useState("alle");
-  const gridRef = useRef<HTMLDivElement>(null);
-  const filteredItems = activeFilter === "alle" ? verstehenInfografiken : verstehenInfografiken.filter(i => i.category === activeFilter);
-
   return (
     <Layout>
       <SEO title="Borderline verstehen" description="Was ist Borderline-Persönlichkeitsstörung? Symptome, Ursachen und Auswirkungen verständlich erklärt." path="/verstehen" />
@@ -542,78 +475,7 @@ export default function Verstehen() {
             </ContentSection>
 
             {/* Materialien zum Download – Kategorie 1: Verstehen */}
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              viewport={{ once: true }}
-              className="mb-12 wave-divider-top"
-              id="materialien"
-              style={{ '--wave-color': 'var(--background)' } as React.CSSProperties}
-            >
-              <h2 className="text-2xl md:text-3xl font-semibold text-foreground mb-6 flex items-center gap-3">
-                <Download className="w-8 h-8 text-sage-dark" />
-                Infografiken zum Thema
-              </h2>
-              
-              <p className="text-muted-foreground mb-6">
-                Vorschau = Web-Bild. «PDF öffnen» öffnet die A4-Druckversion im neuen Tab – Download im PDF-Viewer oben rechts.
-              </p>
-
-              {/* Filter Tabs */}
-              <div className="flex flex-wrap gap-2 mb-6">
-                {verstehenCategories.map((cat) => (
-                  <Button
-                    key={cat.id}
-                    variant={activeFilter === cat.id ? "default" : "outline"}
-                    size="sm"
-                    onClick={() => {
-                      setActiveFilter(cat.id);
-                      setTimeout(() => {
-                        gridRef.current?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-                      }, 50);
-                    }}
-                    className={`whitespace-nowrap shrink-0 ${activeFilter === cat.id ? "bg-sage-dark hover:bg-sage-mid text-white" : ""}`}
-                  >
-                    <cat.icon className="w-4 h-4 mr-1.5" />
-                    {cat.label} ({cat.count})
-                  </Button>
-                ))}
-              </div>
-              
-              <div ref={gridRef} className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                {filteredItems.map((item) => (
-                  <Card key={item.id} className={`${item.featured && activeFilter === "alle" ? "md:col-span-2" : ""} overflow-hidden hover:shadow-lg transition-shadow`}>
-                    <button type="button" className="aspect-[3/4] bg-muted cursor-pointer w-full" onClick={() => window.open(item.webpUrl, '_blank')} aria-label={`${item.alt} – Vorschau öffnen`}>
-                      <img
-                        src={item.webpUrl}
-                        alt={item.alt}
-                        className="w-full h-full object-cover object-top"
-                        loading="lazy"
-                        width={400}
-                        height={223}
-                        decoding="async"
-                      />
-                    </button>
-                    <CardContent className="p-4">
-                      <h3 className="font-medium text-foreground mb-1">{item.title}</h3>
-                      <p className="text-sm text-muted-foreground mb-3">{item.description}</p>
-                      <a href={item.pdfUrl} target="_blank" rel="noopener noreferrer" aria-label={`PDF öffnen: ${item.title} (neuer Tab)`} className="inline-flex items-center justify-center gap-2 rounded-md text-sm font-medium h-9 px-3 w-full border border-input bg-background hover:bg-accent hover:text-accent-foreground transition-colors">
-                        <ExternalLink className="w-4 h-4" />
-                        PDF öffnen
-                      </a>
-                    </CardContent>
-                  </Card>
-                ))}
-              </div>
-
-              <div className="mt-6 text-center">
-                <Link href="/materialien">
-                  <Button variant="outline">
-                    Alle Materialien anzeigen
-                  </Button>
-                </Link>
-              </div>
-            </motion.div>
+            <VerstehenInfografikenSection />
 
             {/* Hoffnung */}
             <motion.div
